counter-server/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
psycopg2==2.8.6
python-crontab==2.5.1
django-cors-headers==3.5.0